Movies Library/requirements.txt
requests
cowsay
tabulate
python-dotenv
pytest
pytest-mock
